moved kdtree class to utilities package
